feat(structure): 구조변경
구조변경
main, detail -> policy 통합

2022.05.23 복세편살
commit by zzpjw
