+++ API/mark/controllers/mark.controller.js
@@ -1,5 +1,5 @@
 const { showMark, pushMark, showMarkRedis, pushMarkRedis, dataCheck, deleteMark, deleteMarkRedis } = require("../services/mark.service")
-const { checkBookmark } = require("../../policies/services/detail.service")
+const { checkBookmark } = require("../../policies/services/policy.service")
 
 exports.getMarks = async (req, res, next) => {
     /*========================================================================================================
