+++ API/search/controllers/search.controller.js
@@ -1,5 +1,5 @@
-const { logic } = require("../../../main/controllers/main.controller")
-const { checkUser, checkData } = require("../../../main/services/main.service")
+const { logic } = require("../../policies/controllers/policy.controller")
+const { checkUser, checkData } = require("../../policies/services/policy.service")
 exports.postSearch = async (req, res, next) => {
     /*========================================================================================================
     #swagger.tags = ['Search']
