--- main/index.js
@@ -1,7 +1,0 @@
-const router = require("express").Router()
-const mainController = require("./controllers/main.controller")
-const authMiddleware = require("../middlewares/auth/auth.middleware")
-
-router.get("/:userId", authMiddleware, mainController.getMain)
-
-module.exports = router
